DSPTables: Use std::array instead of C arrays

+++ Source/Core/Core/DSP/DSPAssembler.h
@@ -5,6 +5,7 @@
 
 #pragma once
 
+#include <cstddef>
 #include <map>
 #include <string>
 
@@ -90,8 +91,8 @@ private:
 
   char* FindBrackets(char* src, char* dst);
   const opc_t* FindOpcode(const char* opcode, u32 par_count, const opc_t* const opcod,
-                          int opcod_size);
-  bool VerifyParams(const opc_t* opc, param_t* par, int count, bool ext = false);
+                          size_t opcod_size);
+  bool VerifyParams(const opc_t* opc, param_t* par, size_t count, bool ext = false);
   void BuildCode(const opc_t* opc, param_t* par, u32 par_count, u16* outbuf);
 
   char* gdg_buffer;
